lexing done

+++ src/parsing/tokenize_utils.c
@@ -6,13 +6,13 @@
 /*   By: hyoyoon <hyoyoon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/09 14:10:00 by hyoyoon           #+#    #+#             */
-/*   Updated: 2024/09/09 14:18:47 by hyoyoon          ###   ########.fr       */
+/*   Updated: 2024/09/09 14:30:36 by hyoyoon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	process_operator_out(char c, int *len, t_token_type *token_type)
+void	process_operator_out(char c, size_t *len, t_token_type *token_type)
 {
 	if (c == '>')
 	{
@@ -26,7 +26,7 @@ void	process_operator_out(char c, int *len, t_token_type *token_type)
 	}
 }
 
-void	process_operator_in(char c, int *len, t_token_type *token_type)
+void	process_operator_in(char c, size_t *len, t_token_type *token_type)
 {
 	if (c == '<')
 	{
